refactor(tui): extract startup guide flow into component
